+++ packages/react-native-editor/CHANGELOG.md
@@ -14,7 +14,8 @@ For each user feature we should also add a importance categorization label  to i
 -   [*] Fixes color picker rendering bug when scrolling [#30994]
 -   [*] Add enableCaching param to fetch request on Android [#31186]
 -   [*] Disabled featured image banner on iOS. [#31681]
--   [*] Enable reusable block only in WP.com sites [#31744]
+-   [***] Add reusable blocks to the inserter menu. [#28495]
+-   [***] Enable reusable block only in WP.com sites [#31744]
 
 ## 1.52.1